move trim_small_chains to backend
